@@ -1,21 +1,15 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Runtime.InteropServices;
-using System.Text;
 
-namespace DungeonRunnersUnpacker.Models
+namespace DungeonRunnersUnpacker.Models;
+
+[StructLayout(LayoutKind.Sequential, Pack = 4)]
+internal struct IndexHeader
 {
-    [StructLayout(LayoutKind.Sequential)]
-    struct IndexHeader
-    {
-        public int Version;
-        public int Unknown1;
-        public int Unknown2;
-        public int Unknown3;
-        public int Unknown4;
-        public int NumInfos;
-        public int Unknown5;
-        public int StringTableSize;
-        public int StringTableOffset;
-    }
+    public int Version;
+    public Guid ArchiveGuid; // pairs index to pkg
+    public int NumInfos;
+    public int DataOffset; // ?
+    public int StringTableSize;
+    public int StringTableOffset;
 }
