@@ -1,34 +1,33 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.IO;
-using System.Text;
 
-namespace DungeonRunnersUnpacker
+namespace DungeonRunnersUnpacker;
+
+internal static class Program
 {
-    class Program
+    private static void Main(string[] args)
     {
-        static void Main(string[] args)
+        var directory = GetDirectory(args);
+        var output = Path.Combine(directory, "dump");
+        var indexFile = new IndexFile(directory);
+        using ArchiveFile archiveFile = new(directory);
+
+        if (indexFile.Header.ArchiveGuid != archiveFile.Guid)
+            throw new Exception("Archive and Index hashes do not match");
+
+        Directory.CreateDirectory(output);
+
+        foreach (var entry in indexFile.Entries)
         {
-            var directory = GetDirectory(args);
-            var output = Path.Combine(directory, "dump");
-            var indexFile = new IndexFile(directory);
-            using var archiveFile = new ArchiveFile(directory);
+            var filename = entry.GetFileName(indexFile);
+            Console.WriteLine(filename);
 
-            Directory.CreateDirectory(output);
-
-            foreach (var entry in indexFile.Entries)
-            {
-                var filename = entry.GetFileName(indexFile);
-                Console.WriteLine(filename);
-
-                archiveFile.ExtractFile(entry, Path.Combine(output, filename));
-            }
+            archiveFile.ExtractFile(entry, Path.Combine(output, filename));
         }
+    }
 
-        static string GetDirectory(string[] args)
-        {
-            return args?.Length == 1 ? args[0] : Directory.GetCurrentDirectory();
-        }
+    private static string GetDirectory(string[] args)
+    {
+        return args?.Length == 1 ? args[0] : Directory.GetCurrentDirectory();
     }
 }
