@@ -1,33 +1,28 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace DungeonRunnersUnpacker.Models;
 
-namespace DungeonRunnersUnpacker.Models
+internal enum FileType
 {
-    enum FileType
-    {
-        tga,
-        bmp,
-        dds,
-        ttf,
-        cobj,
-        d3d,
-        dre,
-        wav,
-        ogg,
-        win,
-        msc,
-        mig,
-        dict,
-        gc,
-        tile,
-        world,
-        zone,
-        dat,
-        fx,
-        cfg,
-        txt,
+    tga,
+    bmp,
+    dds,
+    ttf,
+    cobj,
+    d3d,
+    dre,
+    wav,
+    ogg,
+    win,
+    msc,
+    mig,
+    dict,
+    gc,
+    tile,
+    world,
+    zone,
+    dat,
+    fx,
+    cfg,
+    txt,
 
-        Max = 0x15
-    }
+    Max = 0x15
 }
